pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "The client for Nanoko"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nanoko-python>=0.2.0",
    "pyqt6>=6.9.0",
    "pyqt6-fluent-widgets>=1.8.1",
    "keyring>=24.0.0",
    "matplotlib>=3.7.0",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "nanoko-python",
]

[tool.uv.sources]
nanoko-python = { path = "../nanoko-python", editable = true }
